fluoroview/requirements.txt
numpy>=1.24,<2
tifffile>=2023.7
Pillow>=10.0
scipy>=1.11
scikit-image>=0.21
scikit-learn>=1.3
matplotlib>=3.7
opencv-python-headless>=4.8
cellpose>=3.0
